refactor i18n
