+++ package.json
@@ -44,7 +44,6 @@
     "unplugin-auto-import": "^0.15.3",
     "unplugin-vue-components": "^0.24.1",
     "vite": "^5.0.6",
-    "vite-plugin-json5": "^1.1.2",
     "vitest": "^1.6.0"
   }
 }
