+++ vue/src/board/i18n/chart.cn.json
@@ -1,0 +1,64 @@
+{
+  "charts": {
+    "audio": {
+      "error": "图表无法被显示，因为\"{tag}\"是 {type} 类型, 但是音频图并不支持此类型."
+    },
+    "image": {
+      "error": "图表无法被显示，因为\"{tag}\"是 {type} 类型, 但是图片图并不支持此类型."
+    },
+    "line": {
+      "copy": {
+        "apple": "按下CMD+C拷贝此数据",
+        "success": "已复制！",
+        "windows": "按下CTRL+C拷贝此数据"
+      },
+      "error": "图表无法被显示，因为\"{tag}\"是 {type} 类型, 但是折线图并不支持此类型."
+    },
+    "share": {
+      "smooth": {
+        "method": "算法",
+        "smoothing": "平滑"
+      },
+      "step": "步数",
+      "tag": "标签",
+      "time": "时间",
+      "value": "值",
+      "download": {
+        "button": "下载",
+        "png": {
+          "title": "导出为PNG",
+          "name": "名称",
+          "width": "宽度",
+          "height": "高度",
+          "download": "下载图像"
+        }
+      }
+    },
+    "text": {
+      "titles": {
+        "caption": "说明",
+        "count": "字数",
+        "step": "步数",
+        "tag": "标签",
+        "text": "文本"
+      }
+    }
+  },
+  "empty": "缺少实验数据",
+  "error": "此图表无法正确显示",
+  "search": {
+    "placeholder": "搜索图表..."
+  },
+  "label": {
+    "default": "图表",
+    "pinned": "置顶图表",
+    "hidden": "隐藏图表"
+  },
+  "unkonwn": "未知的图表类型: {type}",
+  "zoom": "放大展示",
+  "pin": "置顶",
+  "unpin": "取消置顶",
+  "more": "更多操作",
+  "hide": "隐藏",
+  "unhide": "取消隐藏"
+}
