--- vue/src/board/i18n/en-US/chart.json5
@@ -1,64 +1,0 @@
-{
-  charts: {
-    audio: {
-      error: 'The Chart cannot be displayed because the value of "{tag}" is type {type}, and Audio Chart do not support displaying type {type}.'
-    },
-    image: {
-      error: 'The Chart cannot be displayed because the value of "{tag}" is type {type}, and Image Chart do not support displaying type {type}.'
-    },
-    line: {
-      error: 'The Chart cannot be displayed because the value of "{tag}" is type {type}, and Line Chart do not support displaying type {type}.',
-      copy: {
-        windows: 'Press CTRL+C to copy this data.',
-        apple: 'Press CMD+C to copy this data.',
-        success: 'Copied!'
-      }
-    },
-    text: {
-      titles: {
-        tag: 'tag',
-        step: 'step',
-        caption: 'caption',
-        count: 'number of words',
-        text: 'Text'
-      }
-    },
-    share: {
-      step: 'Step',
-      tag: 'Tag',
-      time: 'Time',
-      value: 'Value',
-      smooth: {
-        smoothing: 'Smoothing',
-        method: 'Method'
-      },
-      download: {
-        button: 'Download',
-        png: {
-          download: 'Download image',
-          height: 'Height',
-          name: 'Name',
-          title: 'Export PNG',
-          width: 'Width'
-        }
-      }
-    }
-  },
-  empty: 'Empty Experiment with No Charts',
-  error: 'This chart is not displayed correctly.',
-  search: {
-    placeholder: 'Search Chart...'
-  },
-  label: {
-    default: 'Charts',
-    pinned: 'Pinned',
-    hidden: 'Hidden'
-  },
-  unkonwn: 'Unknown chart type: {type}',
-  zoom: 'Zoom In',
-  pin: 'Pin',
-  unpin: 'Unpin',
-  more: 'More actions',
-  hide: 'Hide',
-  unhide: 'Unhide'
-}
