--- vue/src/i18n/en-US/common.json5
@@ -1,55 +1,0 @@
-{
-  'config-editor': {
-    button: 'Edit',
-    'not-allowed': 'Not Allowed Since Is Running',
-    save: {
-      default: 'save changes',
-      savimg: 'saving',
-      success: 'Save Successfully'
-    },
-    'sub-title': {
-      experiment: {
-        desc: 'Description',
-        name: 'Experiment name'
-      },
-      project: {
-        desc: 'Description',
-        name: 'Project name'
-      }
-    },
-    title: {
-      experiment: 'Experiment Information',
-      project: 'Project Information'
-    }
-  },
-  copy: 'Copy successful!',
-  delete: {
-    button: 'Delete',
-    content: {
-      experiment: 'Once you delete a experiment, there is no going back. Please be certain.',
-      project: 'Once you delete a project, there is no going back. Please be certain.'
-    },
-    'not-allowed': 'Not Allowed Since Is Running',
-    title: {
-      experiment: 'Delete this experiment',
-      project: 'Delete this project'
-    }
-  },
-  search: {
-    placeholder: 'Search Lab...'
-  },
-  sidebar: {
-    experiments: {
-      running: 'running',
-      title: 'Experiments'
-    },
-    project: {
-      charts: 'Charts',
-      runs: 'Tables',
-      title: 'Project'
-    }
-  },
-  table: {
-    empty: 'Empty Table'
-  }
-}
