+++ vue/src/i18n/en-US/home.json
@@ -1,0 +1,22 @@
+{
+  "overview": {
+    "create": "Created",
+    "latest": "Latest Run",
+    "total": "Total Experiments",
+    "logdir": "Logdir"
+  },
+  "list": {
+    "title": "Experiments",
+    "table": {
+      "header": {
+        "name": "Name",
+        "status": "State",
+        "create": "Created",
+        "duration": "Runtime"
+      }
+    }
+  },
+  "empty": {
+    "table": "There are no experiments, you can create one by importing SwanLog in your code."
+  }
+}
