--- vue/src/i18n/en-US/sider.json5
@@ -1,12 +1,0 @@
-{
-  version: {
-    version: 'Current Version: v{version}, View',
-    updates: 'Release Notes'
-  },
-  nav: {
-    home: 'Overview',
-    search: 'Search Experiments...',
-    experiment: 'Experiment',
-    help: 'Help'
-  }
-}
