--- vue/src/i18n/zh-CN/common.json5
@@ -1,55 +1,0 @@
-{
-  'config-editor': {
-    button: 'Edit',
-    'not-allowed': 'Not Allowed Since Is Running',
-    save: {
-      default: 'save changes',
-      savimg: 'saving',
-      success: 'Save Successfully'
-    },
-    'sub-title': {
-      experiment: {
-        desc: 'Description',
-        name: 'Experiment name'
-      },
-      project: {
-        desc: 'Description',
-        name: 'Project name'
-      }
-    },
-    title: {
-      experiment: 'Modify Experiment Information',
-      project: 'Modify Project Information'
-    }
-  },
-  copy: '复制成功！',
-  delete: {
-    button: 'Delete',
-    content: {
-      experiment: '一旦你删除了一个实验，就没有回头路了。请确认一下。',
-      project: '一旦你删除了一个项目，就没有回头路了。一定要确定。'
-    },
-    'not-allowed': 'Not Allowed Since Is Running',
-    title: {
-      experiment: '删除这个实验',
-      project: '删除这个项目'
-    }
-  },
-  search: {
-    placeholder: 'Search Lab...'
-  },
-  sidebar: {
-    experiments: {
-      running: '运行中',
-      title: '实验'
-    },
-    project: {
-      charts: '图表',
-      runs: '工作区',
-      title: '项目'
-    }
-  },
-  table: {
-    empty: '暂无数据'
-  }
-}
