+++ vue/src/i18n/zh-CN/index.js
@@ -1,10 +1,10 @@
-import sider from './sider.json5'
-import home from './home.json5'
-import experiment from './experiment.json5'
-import moment from './moment.json5'
-import common from './common.json5'
-import error from './error.json5'
-import chart from '@swanlab-vue/board/i18n/zh-CN/chart.json5'
+import chart from '@swanlab-vue/board/i18n/chart.cn.json'
+import sider from './sider.json'
+import home from './home.json'
+import experiment from './experiment.json'
+import moment from './moment.json'
+import common from './common.json'
+import error from './error.json'
 
 export default {
   chart,
