+++ vue/src/i18n/zh-CN/moment.json
@@ -1,0 +1,20 @@
+{
+  "relativeTime": {
+    "future": "%s之前",
+    "past": "%s前",
+    "s": "几秒钟",
+    "ss": "%d秒",
+    "m": "1分钟",
+    "mm": "%d分钟",
+    "h": "1小时",
+    "hh": "%d小时",
+    "d": "1天",
+    "dd": "%d天",
+    "w": "1周",
+    "ww": "%d周",
+    "M": "1个月",
+    "MM": "%d个月",
+    "y": "1年",
+    "yy": "%d年"
+  }
+}
