--- vue/src/i18n/zh-CN/sider.json5
@@ -1,12 +1,0 @@
-{
-  version: {
-    version: '当前版本：v{version}，查看',
-    updates: '更新日志'
-  },
-  nav: {
-    home: '概览',
-    search: '搜索实验...',
-    experiment: '实验',
-    help: '帮助'
-  }
-}
